Separate isochrone into private module
