+++ .pre-commit-config.yaml
@@ -32,7 +32,7 @@ repos:
       - id: text-unicode-replacement-char
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: "v0.0.281"
+    rev: "v0.0.282"
     hooks:
       - id: ruff
         args:
